requirements.txt
# BİTKİ HASTALIĞI TESPİT SİSTEMİ
# Gerekli Python Paketleri
# Python 3.9+ gereklidir

# Core ML Frameworks
numpy==1.24.3
tensorflow==2.15.0

# PyTorch (Mac M1/M2 için)
torch==2.1.0
torchvision==0.16.0

# YOLO
ultralytics==8.0.196

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0

# Streamlit UI
streamlit==1.28.1

# Data Visualization
plotly==5.18.0

# Utilities
pyyaml==6.0.1
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Optional (for testing)
pytest==7.4.3
